♻️ Refactor: 훅 분리
